test: move epic 3 med pri to tests to its respective folder
